anolis: net/smc: Fix fallback count overflow issue
ANBZ: #5791

The overall count of fallback is a u64 type variable, but the detailed
count of each cause is a u16 type variable, which makes the sum of each
cause count unequals to the overall count. So fix this.

TCP fallback                    2320588 (client 0, server 2320588)
  Server
   PEER_NO_SMC                    25827
   NOSMCRDEV                       1001

+++ net/smc/smc_stats.h
@@ -36,7 +36,7 @@ enum {
 
 struct smc_stats_fback {
 	int	fback_code;
-	u16	count;
+	u64	count;
 };
 
 struct smc_stats_rsn {
